Fix call to SolidFire GetVolumeByName
NDVP changes are in progress to simplify and reuse some of the
SolidFire calls.  As a result the sfapi.GetVolumeByName has been
removed and the logic for this has been moved into the storage_driver.

The goal for sfapi is that it will be limited to native SolidFire API
calls only and be completely consumer agnostic.  There's a lot of
changes that we'll want to do here over time, but the first round took
out the GetVolumeByName helper method, that's ok though we have the
functionality in the storage_driver itself and can use that for now.

During the next release we should dump the default SolidFire API
version to 8.0 which will give us things like this natively and much
more efficiently.

Relies on NDVP commit: cf4fbb74aba