Use multiple VAGS for SolidFire
This change modifies the existing Trident VAG usage to allow
specification of multiple VAGs per deployment.  Currently
Trident uses a single default "Trident" VAG and lumps all of
the initiators and volumes in that single Group.  This is fine
until you want more than 64 nodes in your deployment.

Since the VAG has a max of 64 IQN's, we'll extend things by
enabling up to 4 VAGs to be specified in each deployment.  That
means we can now handle up to 256 nodes.

This change will cause Trident to pull the list of VAG ID's from the
conf file, verify that they exist on the Cluster and for each volume
we create, we'll add it to all of the specified VAGs.

We'll update existing Volumes on Trident restart (add them to the VAGs)
which means if you start with less than 4 vags and need to extend and
add more the restart process will pick those changes up and handle them.

We also change the configuration name to AccessGroups to be more
consistent with other products managed by Trident.

NOTE if you are upgrading an existing deployment, remember you need to
add your Trident VAG to the VagIDs list!!

Addresses:  Github Issue #24
Internal Tracking ID: US2573