Hide Trident REST interface from outside the pod
The Trident REST interface has no authentication layer and is
a potential security vulnerability. This patch changes Trident
to only listen to localhost on its REST interface, which will
only respond inside a K8S pod. Interaction with the REST interface,
such as that in tridentctl, may be accomplished with appropriate
authentication controls by using 'kubectl exec'.

The REST interface may optionally be exposed on the pod IP address by
specifying -address "" as an argument to Trident.