Refactor iSCSI device discovery logic
Device addition & removal should be handled individually, rather than
an entire SCSI bus or controller at a time. This commit removes the reliance
on the iSCSI rescan script, and it also notifies the host of impending device
removals. Any operations that take time on the host are retried using an
exponential backoff.