Remove race condition in o.volumes cache
Previously, the o.volumes map would be recreated from scratch to get
the latest volume information from the backend. Instead, we should
update volume information in-place and remove deleted volumes one by one.
Otherwise, the o.volumes map may be in the middle of reconstruction when
another thread references it.