Added support for Adaptive QoS policies that have been pre-created on the storage system, for all the ontap backends.  The Adaptive QoS policy is applied at the volume level for all ontap backends except for the san-economy driver, where it is applied at the LUN level. For the nas-economy driver, this means that all PVCs share the same Adaptive QoS policy group, but this is the best we can do at this point given what is supported by ONTAP.
To use this feature, configure a Trident backend with with the
adaptivePolicyGroupname as part of the backend per the following example:

{
    "version": 1,
    "storageDriverName": "ontap-nas",
    "backendName": "simple-aqos",
    "managementLIF": "ontapsim1",
    "svm": "vs1",
    "username": "admin",
    "password": "netapp123",
    "adaptivePolicyGroupName": "policy1",
}

Note that a cluster scoped account must be used in the driver
configuration in order to use QoS in clustered ONTAP.  Also note
that support for setting the adaptivePolicyGroupName has been
implemented at the backend driver layer, rather than as a Trident
volume defaults setting, so that end users cannot change the
Adaptive QoS policy on a per PVC basis.  The idea is that Adaptive
QoS should be defined and enforced by the storage administrator and
the Kubernetes administrator, not by the end user.