Introduce Trident Operator to manage Trident's lifecycle
This PR introduces Trident Operator for the Trident Orchestrator.

Developers:

    Rohit Arora
    Clinton Knight

The goal of the Trident operator is to manage the lifecycle of Trident which includes Installing, Uninstalling, Updating, Patching Trident, this is done by the virtue of reconciliation which brings the current state of the Trident installation to the desired state. The desired state of the Trident installation is identified on the basis of the Spec fo the Trident Provisioner CR.

In order to maintain Trident objects Operator needs to discover them and the discovery of Trident objects (deployment, daemonset, service account, pod security policy etc.) is done using labels, as part of the Operator initiative everything created via Trident Operator (Except for the CRDs) is going to be labelled. So, it is very important to have only one instance of the object of a particular kind (like deployment, daemonset etc) with that label, if more than one object of a kind with the same Trident label exist then it should be resolved via the Operator.

Supportability:

    Trident CSI only
    Kubernetes 1.14 to 1.18
    Trident 20.04.x
    Greenfield use cases only.