Add note on behavior of the solidfire-san driver
Element DeleteVolume marks an active volume for deletion. It is purged (permanently deleted) after the cleanup interval elapses.
Here it's purged instantly, which is useful for environments where many volumes get created and deleted, but not the standard behavior.
Perhaps a similar note should be added to the K8s section but I don't have a K8s cluster to verify it.