Add waitForDevice call before calling the blkid
This PR addresses the issue of `blkid` being called in the event of device unavailability, as a result `blkid could throw exit status 2 error. Exit status 2 can be a result of an unformatted device or specified token not found.

So, with the addition of `waitForDevice` call before calling the `blkid` would ensure the device availability when the `blkid` is called.

Also as part of this PR, we are not allowing `fstype` discovery for the raw block volumes and would throw an error if any error other than exit status 2 is identified as a result of the `blkid` call.