Safety check to ensure an unformatted device
This PR adds an extra step in identifying an unformatted device. Today we use `blkid` to identify an unformatted device, blkid returns an exit status=2, which could either mean an unformatted device or device unavailable.  If that is the case, we are also going to read first 2MBs of the device as a safety check to ensure it is, in fact, an unformatted device, if not we would throw an error and Kubernetes would retry the iSCSI volume staging process.