Ensure length of untrimmed "dd" bytes is valid
In order to ensure we have an unformatted device, the 2MiBs of the device are read using "dd" to verify it contains all zeros. It is ensured that the device contains all zeros by trimming all the zeros and leftover length of the slice is zero. However, it is possible that "dd" may read 0 bytes and Trident may misidentify the device as unformatted. It is therefore important to count the bytes read before trimming all the zeros.