Redact sensative information in Trident logs for Sofidfire and E-Series
This PR addresses the issue of sensitive information exposure in the Trident logs. By implementing the stringer interface for the drivers in a manner that is developer-friendly and easy to maintain provides an easy solution to hide any sensitive information in the Trident logs and other places where backend are printed.