Use distroless base image
Alpine provides relatively little value, but carries with it a large
amount of security risks. Despite the team's belief that most of the
security risks don't put users in any actual danger, it is a burden
to evalualte them all. Switching to a base image with a smaller attack
surfaces greatly reduces this burden and gives users more assurance
of security.

As part of this change, CGO is disabled. We didn't need it, and it
was adding needless dependencies.

Also, because this change removes the shell from the container, it
replaces the chroot-wrapper script with a binary that does the same
thing.

Co-authored-by: Andrew Kerr <andrew.kerr@netapp.com>