Prevent delete detach race
* Catch and propogate errors on detach

Swallowing errors on detach was leading to data loss
because we would continue disconnecting even when there
was buffered data that hadn't been flushed to the storage
controller. We now fail at the first sign of problems, and
rely on higher level retries to complete the process.

For Docker mode, the old behavior is preserved, due to
Docker's tight timeout window and lack of retry behavior.
A command line flag is added for the CSI plugin to revert
to the legacy behavior if desired.

* Verify LUN serial numbers before using the LUNs

Pass the correct serial number from the controller to the
node as part of the publish info, then verify the serial
that the node sees. If the serial number is wrong, try to
get the kernel to update it, but be sure to fail if we
cannot update it.

* Improve block device flushing and update sidecar

Update external-provisioner sidecar to v2.0.1.

Use ioctl instead of blockdev binary to flush block
devices.