Fix IPv6 handling during iSCSI attach call
This PR addresses the issue of incorrect parsing of IPv6 when trying to identify if IPv6 portals have been logged in or not during the iSCSI attach call. To fix this issue, Trident is ensuring the IPv6 addresses are contained within square brackets and are properly formatted. This is applicable to not just target portals (like previously) but also to the other portals as well.