Add ability to restrict capacity pools selection in the ANF backend
This PR adds a new field in the ANF backend capacityPools (a list field) that lets users specify the capacity pool they want to use for volume provisioning. This field works can be overridden at the virtual pool's level.

If using this field the other capacity pools that do not match the names of the names in the `capacityPools` list are not considered for volume provisioning purposes.