Modifying parameters in the last version docs
Hello team.

I'm requesting these changes as our customers are getting confusing.

For the StorageClasses: We're not using anymore the provisioner "netapp.io/trident" and needs to be replaced by "csi.trident.netapp.io" in all of them.
For the CVS backend configuration doesn't make sense to have the parameter "limitVolumeSize": "50Gi" since the minimum volume size is 100Gi so using that they will get the error that the size requested is bigger than the one set here.

I really appreciate you take care of this. Thanks a lot !