EOF returned on re-import of a volume
Trying to import an already imported volume creates a PVC successfully. However, we would wait on trying to create the PV and return an error which would end up being an EOF to the Trident REST API invocation.
In the mean time, the new PVC would keep triggering CreateVolumes and erroring saying the PV is already imported.

The fix checks if the volume is already imported and errors out before the PVC is created.