Use fully qualified domain image names by default (docker.io)
It is more and more common that Container Runtimes do not resolve
unqualified image names to docker.io for various reason.

Setting docker.io/ prefix by default does not change the behavior at all
but makes the life much easier for teams managing trident on clusters
without unqualified image name resolution. This would also prevent
deployment errors if an administrator does not pay attention to the
image used in the manifests or in helm.