Allow operator to set PSA in Rancher
In Rancher, it is not enough to have `patch` permissions for a namespace
in order to set PSA labels.
It is also required to have the `updatepsa` permission on the `projects`
resource, as outlined
[here](https://github.com/rancher/rancher/issues/41191).

This rule allows the Trident operator to set the PSA label
`pod-security.kubernetes.io/enforce: privileged` on its installation
namespace in Rancher.

Closes #839