Add the ability to configure and disable iSCSI self-healing
Add the ability to configure and disable iSCSI self-healing via tridentctl, Operator and Helm.

iSCSI Self-Healing Interval: Changing this value influences at what interval iSCSI Self-healing is run (default 5 mins). A user may configure it to run more often by setting a lower number or less frequently by configuring it to a larger value. Setting this to 0 stops iSCSI self-healing completely.

iSCSI Self-Healing Wait Time: Changing this value influences how much time iSCSI self-healing waits before logging out of an unhealthy session and trying to log in again (default 7 mins). A user may configure it to a larger value so that sessions that are identified as unhealthy have to wait longer before being logged out and then an attempt is made to log in again or a smaller value to log out and log in earlier.


thank you...signing off.